+++ src/system/ModuleManager.ts
@@ -2,6 +2,7 @@ import { ModuleInstance } from "@/system/module/ModuleInstance";
 import { debug, err } from "@/system/Console";
 import type { BCPlus } from "@/index";
 import Core from "@/modules/Core";
+import { GUI } from "@/modules/GUI";
 
 export default class ModuleManager {
 
@@ -10,6 +11,7 @@ export default class ModuleManager {
 
     constructor(private readonly parent: BCPlus) {
         this.modules = [
+            new GUI(parent),
             new Core(parent),
         ];
     }